Update useCaptureStep Hook
Hook useCaptureStep now receive getCaptureStepSpec and getRecordStateSpec as parameter
Create a specific hook for DocumentVideoCapture
